backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.2
pydantic-settings==2.3.1
psycopg2-binary==2.9.9
alembic==1.13.1
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
